Cleaned up Trader class
Added a function to calculate the current market value - put this in the generate_bid_ask() and update_price_history() functions.

Cleaned up the submit_shout() function.

Created a create_order() function to reduce some of the redundancy present in the trading_strategy() function -> This function has replaced the code in the aggressive, passive and random branches of the trading strategy.