Fixed the momentum traders
Removed the existing momentum trader implementation and created price action momentum traders, which follow a 25 day high or low.

Added the price history update to the simulator so that each trader now has an accurate trade history for the last 25 'days'.

Updated the orderbook so that it clears out the entire order book, not just when there are overlapping orders.