Added a Trader Type Experiment Script
Reworked the simulator to assign traders their type from a distribution based off the weights assigned to each trader type.

Then added a basic experimentation script for examining each trading type in isolation by systematically increasing the amount of the trader population which abides by that given trading type. Functionality to store the results of each experiment in a text file (in csv format) in a results folder was created.