Trader Type Experiments
Added an experiment script that runs simulations that examine the impact of each implemented trading type.

This is done by systematically increasing the percentage of the population that ascribes to a certain strategy.

The results for the general market and the performance of each trading strategy are then stored for each of these iterations in a corresponding text file in csv-ish format.